packages/jdf-hooks/pyproject.toml
[project]
name = "jdf-hooks"
version = "5.0.0"
description = "Interactive CLI to set up JDF hooks for any project"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "João Dinis Ferreira", email = "hello@joaof.eu" }]
keywords = ["pre-commit", "lefthook", "git", "hooks", "linting", "formatting"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Version Control :: Git",
]

[project.urls]
Homepage = "https://github.com/joaodinissf/jdf-hooks"
Repository = "https://github.com/joaodinissf/jdf-hooks"
Documentation = "https://github.com/joaodinissf/jdf-hooks#readme"

[project.scripts]
jdf-hooks = "jdf_hooks.cli:main"

[project.optional-dependencies]
dev = ["pytest", "pyyaml", "ruff", "pyright"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jdf_hooks"]

[tool.hatch.build.targets.sdist]
include = ["/src"]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["tests/integration/example_files"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # Pyflakes
  "W",   # pycodestyle warnings
  "I",   # isort
  "N",   # pep8-naming
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "SIM", # flake8-simplify
  "PLC", # Pylint Convention
  "PLE", # Pylint Error
  "PLR", # Pylint Refactor
  "PLW", # Pylint Warning
]

ignore = [
  "PLR0913", # Too many arguments - sometimes necessary
  "PLR2004", # Magic value comparison - sometimes clearer
]

[tool.ruff.lint.pylint]
max-args = 8
max-branches = 15
max-returns = 6
max-statements = 50

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
exclude = ["tests/integration/example_files"]

[tool.ty.src]
exclude = ["tests/integration/example_files/"]
